removed build batch files
