+++ .gitignore
@@ -3,3 +3,6 @@
 ################################################################################
 
 /.vs
+/AjaxPro/.vs/AjaxPro
+/AjaxPro/bin
+/AjaxPro/obj
