+++ AjaxPro/AjaxPro.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.1705
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AjaxPro", "AjaxPro.csproj", "{9AD42568-07A4-4D8B-9C6D-1FD54683EF4B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9AD42568-07A4-4D8B-9C6D-1FD54683EF4B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9AD42568-07A4-4D8B-9C6D-1FD54683EF4B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9AD42568-07A4-4D8B-9C6D-1FD54683EF4B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9AD42568-07A4-4D8B-9C6D-1FD54683EF4B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {DCFBFD41-C31A-4F6E-92FC-D451B4918EA5}
+	EndGlobalSection
+EndGlobal
