+++ AjaxPro/Utilities/Constant.cs
@@ -52,6 +52,6 @@ namespace AjaxPro
         /// <summary>
 		/// The assembly version.
 		/// </summary>
-        public const string AssemblyVersion = "9.2.17.1";
+        public const string AssemblyVersion = "21.10.26.1";
 	}
 }
