--- AjaxPro/build_json.bat
@@ -1,6 +1,0 @@
-del release\AjaxPro.JSON.dll
-del release\AjaxPro.JSON.2.dll
-
-"%NET20%\csc.exe" %ARG% /out:"release\AjaxPro.JSON.2.dll" /target:library /define:"TRACE;NET20;NET20external;JSONLIB;%DEFINE%" /r:"System.dll" /r:"System.Data.dll" /r:"System.Drawing.dll" /r:"System.Web.dll" /r:"System.Web.Services.dll" /r:"System.Xml.dll" /r:"System.Configuration.dll" "AssemblyInfo.cs" "Attributes\AjaxNoTypeUsageAttribute.cs" "Attributes\AjaxNonSerializableAttribute.cs" "Attributes\AjaxPropertyAttribute.cs" "Attributes\JavaScriptConverterAttribute.cs" "Attributes\__AjaxClassAttribute.cs" "Attributes\__AjaxEnumAttribute.cs" "Configuration\AjaxSettingsSectionHandler.cs" "JSON\Converters\BitmapConverter.cs" "JSON\Converters\DecimalConverter.cs" "JSON\Converters\EnumConverter.cs" "JSON\Converters\ExceptionConverter.cs" "JSON\Converters\GuidConverter.cs" "JSON\Converters\PrimitiveConverter.cs" "JSON\Converters\StringConverter.cs" "JSON\Converters\DataRowConverter.cs" "JSON\Converters\DataRowViewConverter.cs" "JSON\Converters\DataSetConverter.cs" "JSON\Converters\DataTableConverter.cs" "JSON\Converters\DataViewConverter.cs" "JSON\Converters\DateTimeConverter.cs" "JSON\Converters\HashtableConverter.cs" "JSON\Converters\HtmlControlConverter.cs" "JSON\Converters\JavaScriptObjectConverter.cs" "JSON\Converters\IDictionaryConverter.cs" "JSON\Converters\IEnumerableConverter.cs" "JSON\Converters\IListConverter.cs" "JSON\Converters\NameValueCollectionConverter.cs" "JSON\Converters\ProfileBaseConverter.cs" "JSON\Interfaces\IJavaScriptConverter.cs" "JSON\Interfaces\IJavaScriptObject.cs" "JSON\JavaScriptDeserializer.cs" "JSON\JavaScriptObjects\JavaScriptArray.cs" "JSON\JavaScriptObjects\JavaScriptBoolean.cs" "JSON\JavaScriptObjects\JavaScriptNumber.cs" "JSON\JavaScriptObjects\JavaScriptObject.cs" "JSON\JavaScriptObjects\JavaScriptString.cs" "JSON\JavaScriptObjects\JavaScriptSource.cs" "JSON\JavaScriptSerializer.cs" "JSON\JavaScriptUtil.cs" "JSON\JSONParser.cs" "Utilities\AjaxSettings.cs" "Utilities\Constant.cs" "Utilities\MD5Helper.cs" "Utilities\Utility.cs"
-
-"%NET11%\csc.exe" %ARG% /out:"release\AjaxPro.JSON.dll" /target:library /define:"TRACE;JSONLIB;%DEFINE%" /r:"System.dll" /r:"System.Data.dll" /r:"System.Drawing.dll" /r:"System.Web.dll" /r:"System.Web.Services.dll" /r:"System.Xml.dll" "AssemblyInfo.cs" "Attributes\AjaxNoTypeUsageAttribute.cs" "Attributes\AjaxNonSerializableAttribute.cs" "Attributes\AjaxPropertyAttribute.cs" "Attributes\JavaScriptConverterAttribute.cs" "Attributes\__AjaxClassAttribute.cs" "Attributes\__AjaxEnumAttribute.cs" "Configuration\AjaxSettingsSectionHandler.cs" "JSON\Converters\BitmapConverter.cs" "JSON\Converters\DecimalConverter.cs" "JSON\Converters\EnumConverter.cs" "JSON\Converters\ExceptionConverter.cs" "JSON\Converters\GuidConverter.cs" "JSON\Converters\PrimitiveConverter.cs" "JSON\Converters\StringConverter.cs" "JSON\Converters\DataRowConverter.cs" "JSON\Converters\DataRowViewConverter.cs" "JSON\Converters\DataSetConverter.cs" "JSON\Converters\DataTableConverter.cs" "JSON\Converters\DataViewConverter.cs" "JSON\Converters\DateTimeConverter.cs" "JSON\Converters\HashtableConverter.cs" "JSON\Converters\HtmlControlConverter.cs" "JSON\Converters\JavaScriptObjectConverter.cs" "JSON\Converters\IDictionaryConverter.cs" "JSON\Converters\IEnumerableConverter.cs" "JSON\Converters\IListConverter.cs" "JSON\Converters\NameValueCollectionConverter.cs" "JSON\Interfaces\IJavaScriptConverter.cs" "JSON\Interfaces\IJavaScriptObject.cs" "JSON\JavaScriptConverterList.cs" "JSON\JavaScriptDeserializer.cs" "JSON\JavaScriptObjects\JavaScriptArray.cs" "JSON\JavaScriptObjects\JavaScriptBoolean.cs" "JSON\JavaScriptObjects\JavaScriptNumber.cs" "JSON\JavaScriptObjects\JavaScriptObject.cs" "JSON\JavaScriptObjects\JavaScriptString.cs" "JSON\JavaScriptObjects\JavaScriptSource.cs" "JSON\JavaScriptSerializer.cs" "JSON\JavaScriptUtil.cs" "JSON\JSONParser.cs" "Utilities\AjaxSettings.cs" "Utilities\Constant.cs" "Utilities\MD5Helper.cs" "Utilities\Utility.cs"
