+++ README.md
@@ -67,3 +67,10 @@ function getServerTime_callback(res) {
   alert(res.value);
 }
 ```
+
+# Compiler Options
+
+- `NET20` compiles .NET 2.0 assemblies AjaxPro.2.dll (otherwise original it was .NET 1.1, AjaxPro.dll)
+- `JSONLIB` compiles JSON parser only (AjaxPro.JSON.2.dll or AjaxPro.JSON.dll)
+- `NET20external` is setting the assembly name to AjaxPro.2.dll, compatibility
+- `TRACE` is no longer used
